Break up Lang and Runtime types files
