+++ src/Pact/Docgen.hs
@@ -23,7 +23,6 @@ import Control.Monad.Catch
 import Data.Monoid
 
 import Pact.Types.Lang
-import Pact.Types.Util
 import Pact.Native
 import Pact.Repl
 import Pact.Repl.Lib
